fit_param printed

+++ mocap_bridge/include/udp_msg.h
@@ -39,7 +39,7 @@ struct {
  */
 struct {
     stamp_t stamp;
-    float fit_param;
+    double fit_param;
     odom_t ekf;
     bool sensor_status[10];
 } typedef udp_msg_t;
